Optimized assets
Lossless compression made with the Leanify tool.
Copy of the original mod icon is available on the repository root and the one included in the jar is also resized because the dimensions were an overkill (Mod Menu icon box is kinda small) and the Fabric project recommends dimensions of the power of the 2.
Also, added the issues tag to the fabric.mod.json file.